+++ src/BobCrm.App/Models/Widgets/NumberWidget.cs
@@ -110,4 +110,9 @@ public class NumberWidget : TextWidget
         builder.CloseElement();
         builder.CloseElement();
     }
+
+    public override string GetDefaultCodePrefix()
+    {
+        return "number";
+    }
 }