+++ src/BobCrm.App/Models/Widgets/SelectWidget.cs
@@ -111,4 +111,9 @@ public class SelectWidget : TextWidget
         builder.CloseElement();
         builder.CloseElement();
     }
+
+    public override string GetDefaultCodePrefix()
+    {
+        return "select";
+    }
 }